Query GraphCMS with Appollo Client for new Blog component
Appollo Client allows the frontend to grab content from GraphQL,
in this case GraphCMS from our project. The project has a URL for
grabbing data from GraphCMS publicly, as it is in `index.js`. The
`App` is then wrapped there in a provider HOC for the client.

Other files, IE `BlogPosts FC` (FC: functional component), can query
GraphQL and it recognizes Apollo.

I was stuck for an embarrassing time on mapping all tags for a post
inside the mapping of post data used to list each post on the blog's
hoome page. Thank you to RandellDawson of the freeCodeCamp.org staff
for explaining how that is done in this thread:
https://forum.freecodecamp.org/t/mapping-an-array-within-a-mapped-array/229335/14