Rename all workshop components and import all workshop preview data
Mess of a commit, but mostly because file hierarchy was arranged
inconistently and with incorrect names.

This was intended to allow seeing all of the sections of a workshop
from a page of the workshop, as well as the titles of all of the
lessons under each section title. However, the component names were
misleading.

WorkshopSection was turned to Workshop, and Workshop's export
was named from the MUI doc I borrowed the starter from. The
WorkshopHome no longer iterates over Workshop's but instead WorkshopPreviews.
Workshops embeds a WorkshopSections component that iterates over all
sections of the workshop. WorkshopSections embeds a WorkshopLessons
component that iterates over all lessons in that workshop.

Arguably, the convention of the WorkshopHome iterating over previews
vs the other Workshop components linking to one component that has
its own iterating is inconsistent. I need to make this consistent
when I actually link the previews to workshops (one workshop is
hardcoded in and has to be navigated to by URL), or establish a
hierarchy consistency of previews vs actual items.

Addresses #3 (Pull Workshop preview info from GraphCMS)
Addresses #4 (Pull Lesson data from GraphCMS)
Addresses #5 (Pull Section data from GraphCMS)
Project: 'Create Workshop Lesson Listing'