Retrieve `Workshop FC` via react-router id params
This was an instance of something not being suprising or tricky
once finished, but a few different libraries having conflicting docs.

A lot of docs on react-router's `useParams` were outdated, but this
way of doing so is the current best way. So is the way I can tell of
passing them into the route, which I saw three other alternatives
which were not as simple as this.

Additionally, examples of passing in the `id` variable into the
Apollo Client `useQuery` differed in how the variables were set up.
Again, this was the simplest way in the end.

Lastly, there are differing examples and docs on the type to pass in
for the `id` in the `query GetWorkshop($id: Integer!)`. Some said that
it was type string, which really threw me through a loop when that
was plain wrong. That said, I should have looked at the docs on
system fields, as the note about using `Integer!` were there the
whole time. Of note, type `ID` also works too.

Project: 'Create Workshop Lessons Listing'