Retrieve workshop sections by id, convert id queries to type `ID`
Honestly I am further confused. The only thing in the GraphCMS docs
that uses `ID` is the Users schema which is included by default in
every project. The docs say everything else that is created uses
`Integer!`. This didn't make sense because the items were strings,
but I tried to follow the docs on that one. I must have misread
the docs.

Project: 'Create Workshop Lessons Listing'