Add demo component area for a lesson demo
I am having issues at the moment with routing to the actual page,
`WorkshopLesson FC`, due to how values are held when at a child
route under `/workshops/{workshop}/{lesson}`. So for now, I am
toying with making standalone pages in a dedicated demo location
that could be routed to in launched deployments of dev stagings.

Project: 'Create lesson page'