Change state of lesson completion in `CompleteButton sub-FC`
`...sub FC` is what I just decided to be the way to declare that a
component within a component that is exported. So it is a subcomp
of `LessonDemo FC`. To not make the commit header longer, I will
let the context of the surrounding commits speak for the parent
component.

Project: 'Create Lesson page'